% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwhs_palette_package_v03.R
\name{cwhs_pal}
\alias{cwhs_pal}
\title{Function to extract cwhs_online colours as hex codes}
\usage{
cwhs_pal(palette = "online", reverse = FALSE, ...)
}
\arguments{
\item{...}{Character names of cwhs_online colours}
}
\description{
Function to extract cwhs_online colours as hex codes
}
